Update recommendedRules.json
Instagram slightly changed their HTML on video player pages which made this rule set stop working. There's a div that blocks you from clicking on the controls. This selector gets that div by getting the adjacent sibling combinator of the video, and then the div within that div, and sets the pointer events to none, which stops it from blocking you from clicking on the controls.